CVhomework_1/requirements.txt
numpy==1.14.3
matplotlib==2.2.2
opencv_python==4.4.0.44
Pillow==8.0.1
